2025/03/06 -- 2
Added the `squeeze` parameter to set the `squeeze` parameter of the internally called `subplots`. For the sake of existing code, it was decided not to forcibly set `squeeze` to `False`, but to continue maintaining it as `True` (consistent with the original behavior of `subplots`). At the same time, if n_row == n_col == 1, the returned `axes` will retain the previous behavior of being converted into a 1D array, which is different from `subplots`.